Tweak logic and layout
The bindings elements are only used for fields that will be sent
back to the server, this field is purely client-side.

Rather than hiding the box, which makes the page jump, disable it
and clear the value.

Tweaked wording, moved help text outside of form check so that
it aligns with controls no matter different sizes of checkboxes
in different browsers.