Judge completeness of works in a place by breadth completeness
We currently display breadth completeness everywher, ignoring depth
completeness (which requires a point-in-time for every work) for now.

We therefore want to calculate the daily number of complete works
based on breadth completeness, not just expressions.

This corrects, for example, the fact that in za-cpt it shows 59 of 70
works complete, but filtering on incomplete works shows only 4.