Ensure rows have entries for all headers
This fixes an issue where the last text value in a row comes before
the last column, eg. commencement date is filled in, but columns such
as stub come after that. Before this fix, the stub column would
not have been included at all, and so it assumes that 'stub' wasn't
present in the spreadsheet, so falls back to checking for the old
'principal' value.

Now, we correctly see that stub is present and so don't override it.