Helper to support reversing content API URLs.
This is needed for the case when the content API is run on a separate
domain using django_hosts. Sometimes non-API code (such as indigo_app)
needs to reverse indigo_content_api urls. When django_hosts is in use,
this fails, because the API urlconf is not loaded.

With this patch, we detect if django_hosts is in use and inject a
host='api' parameter into the call to DRF's reverse(), AND we ensure that
DRF's reverse() call makes use of django_host's reverse(), which will
load the correct urlconf.