provide for only one commencement relevant to current PiT
I.e. if the work has multiple commencements, but at this point in time there is one relevant commencement and no uncommenced provisions, display the commencement as if it's the only one.
(Otherwise it's confusing to readers to see 'There are multiple commencements' and then a table with only one row.)