add elements to base toc, return parents + children
This is being put aside for now.

We need to consider the implications of changing the
structure of the toc.
(can we adjust and use the regular toc,
or should we generate a special one for commencements only?)

Then we need to work out how to display the nested toc
in the Commencements view, submitting the form, etc.

Once we can record the data, we need to figure out
how to communicate it:
- coverpage
- json

and do the rest (migration, adjust signoff if needed, etc)