re-render only part of the html
this saves us from re-rendering an entire document when quick editing
just a portion

if chp_1__sec_1 is quick edited, then the mutation target will be chp_1
and so all of chp_1 is re-rendered.

We could be smarter and only re-render chp_1__sec_1, but it gets more
complicated. We would then have to detect if the mutation impacts other
siblings (eg. the section was split into two) or if it was deleted.