process mutations in bulk
When eIds are re-written, we can have hundreds of attribute modification
mutations. These are slow when processed individually. Instead, we
process them as a group. After we find first mutation we find in the list that
impacts a component, we can stop processing the remainder.

In the common case, bulk attribute modifications happen after the tree
is modified by a node change; so the first change in the list is a node
change. In many cases this causes the editors to update and so the
attribute modifications are ignored completely.