revert to not using formatting tags when diffing
these get two complex to render, because they can contain nested diffs.

For example, if a <b> tag is changed to <i> AND it has nested text
changed, then we will have a diff with

<b diff:delete-formatting><i diff:insert-formatting>text <ins>inserted</ins></i></b>

and it's unclear how to render that. We'd land up just duplicating the
subtrees as del and ins, which is the same as not comparing them from
the start.