cache Country and Locality lookups, which should reduce db calls
These are used heavily, often in places where prefetch_related or
selected_related is hard to control. Because they don't change
regularly, we can take advantage of caching.

Each process has an LRU cache of size one, keyed on a version number.
When a model is saved/deleted, the cache version key is bumped (which
all processes can see) and so the cache is invalidated and re-fetched.